[refactor] Common component re export
