+++ src/components/Common/Filter/index.ts
@@ -1,0 +1,7 @@
+export { default as AgeFilter } from './AgeFilter';
+export { default as BottomFilter } from './BottomFilter';
+export { default as ColorFilter } from './ColorFilter';
+export { default as Filters } from './Filters';
+export { default as GenderFilter } from './GenderFilter';
+export { default as RangeFilter } from './RangeFilter';
+export { default as StyleFilter } from './StyleFilter';
