+++ src/page/auth/Member.tsx
@@ -1,4 +1,3 @@
-import Logo from 'components/Common/Logo';
 import React from 'react';
 import { ReactComponent as ValidationErrorSvg } from '@svg/validation-error.svg';
 import styled from 'styled-components';
@@ -16,6 +15,7 @@ import useApiNavigation from 'hooks/useApiNavigation';
 import { useAppDispatch } from 'redux/store';
 import { setToken } from 'redux/reducer/auth';
 import { AuthSubmitButton, Input } from 'components/Auth';
+import { Logo } from 'components/Common';
 
 function Member() {
   const {
