+++ src/page/auth/SignUp.tsx
@@ -1,4 +1,3 @@
-import Logo from 'components/Common/Logo';
 import React from 'react';
 import styled from 'styled-components';
 import { ReactComponent as ValidationErrorSvg } from '@svg/validation-error.svg';
@@ -10,6 +9,7 @@ import { useApi } from 'hooks/useApi';
 import { joinByEmail } from 'api/auth';
 import useApiNavigation from 'hooks/useApiNavigation';
 import { AuthSubmitButton, Input } from 'components/Auth';
+import { Logo } from 'components/Common';
 
 function SignUp() {
   const {
